Renamed EventsGrantsPage component to include events"
