+++ apps/charterafrica/src/components/EventsGrantsAndFellowships/index.js
@@ -1,0 +1,3 @@
+import EventsGrantsAndFellowships from "./EventsGrantsAndFellowships";
+
+export default EventsGrantsAndFellowships;
