--- apps/charterafrica/src/components/GrantsAndFellowships/index.js
@@ -1,3 +1,0 @@
-import GrantsAndFellowships from "./GrantsAndFellowships";
-
-export default GrantsAndFellowships;
